pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.1.0",
    "flasgger>=0.9.7.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.39",
    "uwsgi>=2.0.28",
]
